requirements.txt
PyYAML>=6.0
tqdm>=4.66.5